chore: BI-7445 enable ruff RUF046 (unnecessary-cast-to-int) (d392efe28)
Enable the deferred RUF046 rule and apply its fix: drop the redundant
`int()` wrapping a single-argument `round()` — `round(float)` already
returns an `int`, so the cast is a no-op. 4 sites, all `int(round(...))`
in profiling/timing code across mainrepo.

Verified: ruff RUF046 clean across both config trees; `task dev:lint`
green on dl_api_commons, dl_app_tools, dl_core, dl_maintenance
(ruff + mypy confirm the values stay `int`); toml-sort clean on root and
mainrepo pyproject.

Part of BI-7445 (enable deferred `Later` ruff rules); umbrella BI-7321.

+++ lib/dl_api_commons/dl_api_commons/reporting/profiler.py
@@ -134,7 +134,7 @@ class DefaultReportingProfiler(ReportingProfiler):
             # auto: request_id
             **self.rci.get_reporting_extra(),
             connection_type=start_record.connection_type.name,
-            execution_time=int(round((end_timestamp - start_record.timestamp) * 1000)),
+            execution_time=round((end_timestamp - start_record.timestamp) * 1000),
             query=self._obfuscate_query(start_record.query),
             status="success" if error is None else "error",
             error=error_text,
